Day 17 solution
Mathematical solution for part 1 (reasoning is printed) and
brute force for part 2. I appreciate that heuristics
could have been used, e.g. find every feasible x for a y that
works and work out which regions of y are likely to work
given the ymax.
This only adds heuristics based on the best solution and
the limiting areas (getting to ylim/xlim at t=1).