Uploadcommit for Day 22.
Solved part1 naively but obvs. this doesn't scale for part2.
Developed an algorithm adding cubes to the block_list (an 'on' list)
and either splitting the to be added cube, if it's an on cube and overlapping
with one of the cubes in the block_list, or splitting the cube in the
block_list if an 'off' cube which is currently added overlaps with it.
Does work for the simplest input of the example, but underestimates
the number of on cubes for other inputs.